# [perf-tests] make the bucket name and prefix configurable for ray data image resize job